+++ Core/DenLoader.hpp
@@ -27,7 +27,7 @@ namespace DenLoader
     void init();
     Den getDen(u8 index, u8 rarity);
     u8 getLocation(u8 index);
-    QVector<u8> getCoordinates(u8 index);
+    QVector<u16> getCoordinates(u8 index);
 }
 
 #endif // DENLOADER_HPP
